+++ src/MR.Analyzers.Whitespace/TrailingWhitespaceAnalyzer.cs
@@ -20,6 +20,11 @@ namespace MR.Analyzers.Whitespace
 
 		private void AnalyzeSyntaxTree(SyntaxTreeAnalysisContext context)
 		{
+			if (context.Tree.IsGenerated())
+			{
+				return;
+			}
+
 			var root = context.Tree.GetRoot();
 
 			// Trailing whitespace